Move the identity card to shared as CatalogIdentity
The card presents any catalog entity, not only a service account, so it moves to src/components/shared/CatalogIdentity behind a barrel: CatalogIdentity for one that exists, CatalogIdentityEditor for one being named, layout.ts holding the measurements they share, and catalogName.ts holding splitCatalogName and the monogram helpers the tiles are drawn from.

CatalogIdentity now takes the catalog name whole and splits it itself, so a caller cannot pass a leaf and a prefix that disagree. The details page hands it serviceAccount.catalogName and keeps no split of its own.

The monogram falls back to a question mark for a leaf with nothing alphanumeric in it. The old 'SA' fallback named the one feature the card started in. Its comment claimed two letters while the code took one; the comment now says one.

AccountCard, CompactAccountCard, and the details page import the monogram helpers from the barrel. ServiceAccounts/shared.ts keeps what is genuinely about service accounts: capabilities, key lifetimes, and the feature description.
