+++ src/components/shared/CatalogIdentity/index.ts
@@ -1,0 +1,8 @@
+export { CatalogIdentity } from './Display';
+export { CatalogIdentityEditor } from './Edit';
+export {
+    monogram,
+    MONOGRAM_TEXT_COLOR,
+    monogramColor,
+    splitCatalogName,
+} from './catalogName';
